moved: all css and scss files to styles directory
