+++ src/components/projects/index.tsx
@@ -1,4 +1,4 @@
-import './projects.css'
+import '../../styles/projects.css'
 
 export const ProjectsContainer: React.FC = () => {
   return (
