+++ src/components/tech-stack/index.tsx
@@ -1,4 +1,4 @@
-import './tech-stack.css'
+import '../../styles/tech-stack.css'
 
 export const TechStackContainer: React.FC = () => {
   return <section id="tech-stack">TechStack</section>;
