+++ src/styles/bio-container.scss
