+++ src/styles/landing-container.scss
